Publish DEPS for Chromium 48.0.2564.116
Incrementing VERSION to 48.0.2564.116

Cr-Commit-Position: refs/branch-heads/2564@{#706}
Cr-Branched-From: 1283eca15bd9f772387f75241576cde7bdec7f54-refs/heads/master@{#359700}

+++ chrome/browser/ui/webui/engagement/site_engagement_ui.cc
@@ -12,7 +12,7 @@
 #include "content/public/browser/web_ui_data_source.h"
 #include "grit/browser_resources.h"
 #include "mojo/common/url_type_converters.h"
-#include "third_party/mojo/src/mojo/public/cpp/bindings/strong_binding.h"
+#include "mojo/public/cpp/bindings/strong_binding.h"
 
 namespace {
 
@@ -37,7 +37,7 @@ class SiteEngagementUIHandlerImpl : public SiteEngagementUIHandler {
 
     SiteEngagementService* service = SiteEngagementService::Get(profile_);
 
-    for (const std::pair<GURL, int>& info : service->GetScoreMap()) {
+    for (const std::pair<GURL, double>& info : service->GetScoreMap()) {
       SiteEngagementInfoPtr origin_info(SiteEngagementInfo::New());
       origin_info->origin = mojo::String::From(info.first);
       origin_info->score = info.second;
